Ref convert serializer to function
