+++ common/protocol/qconsumerprotocolhandler.go
@@ -3,6 +3,7 @@ package protocol
 import (
 	"github.com/brunograssano/Distribuidos-TP1/common/data_structures"
 	"github.com/brunograssano/Distribuidos-TP1/common/middleware"
+	"github.com/brunograssano/Distribuidos-TP1/common/serializer"
 )
 
 type ConsumerProtocolInterface interface {
@@ -11,16 +12,14 @@ type ConsumerProtocolInterface interface {
 }
 
 type ConsumerQueueProtocolHandler struct {
-	consumer   middleware.ConsumerInterface
-	serializer *data_structures.Serializer
-	recvCount  int
+	consumer  middleware.ConsumerInterface
+	recvCount int
 }
 
 func NewConsumerQueueProtocolHandler(consumer middleware.ConsumerInterface) *ConsumerQueueProtocolHandler {
 	return &ConsumerQueueProtocolHandler{
-		consumer:   consumer,
-		serializer: data_structures.NewSerializer(),
-		recvCount:  0,
+		consumer:  consumer,
+		recvCount: 0,
 	}
 }
 
@@ -29,7 +28,7 @@ func (q *ConsumerQueueProtocolHandler) Pop() (*data_structures.Message, bool) {
 	if !ok {
 		return nil, ok
 	}
-	msg := q.serializer.DeserializeMsg(bytes)
+	msg := serializer.DeserializeMsg(bytes)
 	if msg.TypeMessage == data_structures.FlightRows {
 		q.recvCount += len(msg.DynMaps)
 	}
