+++ common/test/data_structures/dynamicmap_serializer_test.go
@@ -5,6 +5,7 @@ import (
 	"encoding/binary"
 	"fmt"
 	dataStructures "github.com/brunograssano/Distribuidos-TP1/common/data_structures"
+	"github.com/brunograssano/Distribuidos-TP1/common/serializer"
 	"strings"
 	"testing"
 )
@@ -14,7 +15,7 @@ func TestSerializeDynMapWithAnIntReturnsTheBytesExpected(t *testing.T) {
 	dynMap["test"] = make([]byte, 4)
 	binary.BigEndian.PutUint32(dynMap["test"], uint32(32))
 	row := dataStructures.NewDynamicMap(dynMap)
-	serializer := dataStructures.NewSerializer()
+
 	serializedRow := serializer.SerializeDynMap(row)
 
 	var bytesExpected []byte
@@ -48,7 +49,7 @@ func TestSerializeDynMapWithTwoColumnsReturnsTheBytesExpected(t *testing.T) {
 	dynMap["test_string"] = []byte(stringCol2)
 	binary.BigEndian.PutUint32(dynMap["test"], uint32(32))
 	row := dataStructures.NewDynamicMap(dynMap)
-	serializer := dataStructures.NewSerializer()
+
 	serializedRow := serializer.SerializeDynMap(row)
 
 	var bytesExpected []byte
@@ -86,7 +87,6 @@ func TestSerializeDynMapWithTwoColumnsReturnsTheBytesExpected(t *testing.T) {
 
 func TestSerializeEmptyDynMapReturnsOnlyColLengthOfZeroAsBytes(t *testing.T) {
 	row := dataStructures.NewDynamicMap(make(map[string][]byte))
-	serializer := dataStructures.NewSerializer()
 	serializedRow := serializer.SerializeDynMap(row)
 	bytesExpected := make([]byte, 4)
 	binary.BigEndian.PutUint32(bytesExpected, uint32(0))
@@ -129,7 +129,6 @@ func TestDeserializeDynMapWithTwoColumnsReturnsTheExpectedDynMap(t *testing.T) {
 	bytesSer = append(bytesSer, bytesLenValue2...)
 	bytesSer = append(bytesSer, bytesValue2...)
 
-	serializer := dataStructures.NewSerializer()
 	rowReceived, _ := serializer.DeserializeDynMap(bytesSer)
 
 	stringCol, errStringCol := rowReceived.GetAsString("test_string")
