+++ data_processor/dataprocessor_test.go
@@ -51,7 +51,6 @@ func TestShouldGetAMessageProcessItAndSendItToAllChannels(t *testing.T) {
 	outputEx13 := make(chan *dataStructures.Message, 10)
 	outputEx2 := make(chan *dataStructures.Message, 10)
 	outputEx4 := make(chan *dataStructures.Message, 10)
-	serializer := dataStructures.NewSerializer()
 
 	mConsumer := &mockConsumer{
 		inputChannel: input,
@@ -73,7 +72,6 @@ func TestShouldGetAMessageProcessItAndSendItToAllChannels(t *testing.T) {
 		consumer:       mConsumer,
 		producersEx123: []protocol.ProducerProtocolInterface{mProducer2, mProducer13},
 		producersEx4:   mProducer4,
-		serializer:     serializer,
 		ex123Columns:   []string{"startingAirport", "segmentsArrivalAirportCode", "totalStopovers", "route"},
 		ex4Columns:     []string{"route"},
 	}
@@ -124,10 +122,8 @@ func TestShouldGetAMessageProcessItAndSendItToAllChannels(t *testing.T) {
 }
 
 func TestShouldProcessTheDataOfEx123(t *testing.T) {
-	serializer := dataStructures.NewSerializer()
 	processor := &DataProcessor{
 		processorId:  0,
-		serializer:   serializer,
 		ex123Columns: []string{"totalStopovers", "route"},
 	}
 	dynMap := make(map[string][]byte)
@@ -161,10 +157,8 @@ func TestShouldProcessTheDataOfEx123(t *testing.T) {
 }
 
 func TestShouldReturnAnErrorIfTheSegmentsColDoesNotExist(t *testing.T) {
-	serializer := dataStructures.NewSerializer()
 	processor := &DataProcessor{
 		processorId:  0,
-		serializer:   serializer,
 		ex123Columns: []string{"totalStopovers", "route"},
 	}
 	dynMap := make(map[string][]byte)
@@ -179,10 +173,8 @@ func TestShouldReturnAnErrorIfTheSegmentsColDoesNotExist(t *testing.T) {
 }
 
 func TestShouldReturnAnErrorIfTheStartingAirportColDoesNotExist(t *testing.T) {
-	serializer := dataStructures.NewSerializer()
 	processor := &DataProcessor{
 		processorId:  0,
-		serializer:   serializer,
 		ex123Columns: []string{"totalStopovers", "route"},
 	}
 	dynMap := make(map[string][]byte)
